+++ playground/Coalesce.Web.Vue3/src/App.vue
@@ -20,7 +20,8 @@
       <v-btn variant="text" to="/test">Test</v-btn>
       <v-btn variant="text" to="/test-setup">Test2</v-btn>
       <v-btn variant="text" to="/audit-logs">Audit</v-btn>
-      <v-btn variant="text" to="/swagger">OpenAPI</v-btn>
+      <v-btn variant="text" href="/swagger">Swagger</v-btn>
+      <v-btn variant="text" href="/scalar/v1">OpenAPI</v-btn>
       <v-btn variant="text" href="/coalesce-security">Security Overview</v-btn>
 
       <v-menu offset-y>