+++ src/IntelliTect.Coalesce/Api/OpenApi/CoalesceApiDescriptionProvider.cs
@@ -1,15 +1,8 @@
-﻿using IntelliTect.Coalesce.Models;
-using IntelliTect.Coalesce.TypeDefinition;
-using Microsoft.AspNetCore.Mvc;
+﻿using IntelliTect.Coalesce.TypeDefinition;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
-using Microsoft.AspNetCore.Mvc.ModelBinding.Metadata;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace IntelliTect.Coalesce.Api
 {
@@ -17,7 +10,7 @@ namespace IntelliTect.Coalesce.Api
     /// Performs adjustments to the API metadata so that it doesn't cause .NET 9's OpenAPI generation to implode.
     /// In particular, we have to remove the parameters that are bound with Coalesce's custom model binders,
     /// since these cause the OpenAPI generator to throw exceptions. We re-add these definitions with
-    /// CoalesceApiOperationFilter (for Swashbuckle), or CLASS_TBD for the native aspnetcore OpenApi gen.
+    /// CoalesceApiOperationFilter.
     /// </summary>
     internal class CoalesceApiDescriptionProvider(ReflectionRepository reflectionRepository) : IApiDescriptionProvider
     {
@@ -50,7 +43,6 @@ namespace IntelliTect.Coalesce.Api
             // We add them back in in CoalesceApiOperationFilter.
             // They break the new Microsoft.AspNetCore.OpenApi package in .NET 9
             // if we leave them present in the API descriptions.
-
             foreach (var paramVm in method.Parameters.Where(p =>
                 p.Type.IsA(typeof(IBehaviors<>)) || 
                 p.Type.IsA(typeof(IDataSource<>))
@@ -72,12 +64,33 @@ namespace IntelliTect.Coalesce.Api
                     // Remove params that have no setter. 
                     // Honestly I'm clueless why this isn't default behavior, but OK.
                     !p.PropViewModel.HasSetter
-
-                    // Remove the string "DataSource" parameter that is redundant with our data source model binding functionality.
-                    || (p.PropViewModel.Name == nameof(IDataSourceParameters.DataSource) && p.OperationParam.Type == typeof(string))
                 ))
                 {
                     parameters.Remove(noSetterProp.OperationParam);
+                }
+            }
+
+            foreach (var paramVm in method.Parameters.Where(p => p.Type.IsA(typeof(IDataSource<>))))
+            {
+                var declaredFor =
+                    paramVm.GetAttributeValue<DeclaredForAttribute>(a => a.DeclaredFor)
+                    ?? paramVm.Type.GenericArgumentsFor(typeof(IDataSource<>))!.Single();
+
+                var dataSources = declaredFor.ClassViewModel!.ClientDataSources(reflectionRepository);
+
+                foreach (var dataSource in dataSources)
+                {
+                    foreach (var dsParam in dataSource.DataSourceParameters)
+                    {
+                        operation.ParameterDescriptions.Add(new ApiParameterDescription()
+                        {
+                            Source = BindingSource.Query,
+                            Name = $"{paramVm.Name}.{dsParam.Name}",
+                            IsRequired = false,
+                            Type = dsParam.Type.TypeInfo,
+                            ModelMetadata = operation.ParameterDescriptions.First().ModelMetadata.GetMetadataForProperty(dataSource.Type.TypeInfo, dsParam.Name)
+                        });
+                    }
                 }
             }
         }